+++ NEWS.md
@@ -1,5 +1,5 @@
 # tbgeneratr 0.1.2.0
-#### 9 August 2019
+#### 14 August 2019
 ## Major update
 * added function to recalculate treatment outcome for DR-TB in EpiInfo data